geoapi/requirements.txt
Django==4.2
djangorestframework==3.14
psycopg2-binary==2.9.7
gunicorn==20.1.0
djangorestframework-gis==1.1
django-cors-headers==3.14.0